move: Book DTO 위치 변경
