+++ src/main/java/com/nhnacademy/store99/front/book/service/BookService.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.store99.front.book.service;
 
-import com.nhnacademy.store99.front.book.Response.BookPageResponse;
-import com.nhnacademy.store99.front.book.Response.BookResponse;
+import com.nhnacademy.store99.front.book.dto.response.BookPageResponse;
+import com.nhnacademy.store99.front.book.dto.response.BookResponse;
 import com.nhnacademy.store99.front.common.response.CustomPageImpl;
 import org.springframework.data.domain.Pageable;
