+++ src/main/java/com/nhnacademy/store99/front/book/service/BookServiceImpl.java
@@ -1,9 +1,9 @@
 package com.nhnacademy.store99.front.book.service;
 
 
-import com.nhnacademy.store99.front.book.Response.BookPageResponse;
-import com.nhnacademy.store99.front.book.Response.BookResponse;
 import com.nhnacademy.store99.front.book.adapter.BookAdapter;
+import com.nhnacademy.store99.front.book.dto.response.BookPageResponse;
+import com.nhnacademy.store99.front.book.dto.response.BookResponse;
 import com.nhnacademy.store99.front.common.response.CustomPageImpl;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
